\name{bleep}
\alias{bleep}
\title{An R terminal bell}
\description{Beeps by calling a system command called 'bleep'}
\usage{bleep()}
\details{Requires a system function called bleep. For example:

#!/bin/bash

paplay /usr/share/sounds/freedesktop/stereo/complete.oga &

}
\author{
Alexander Hopkins
}
\note{}
