% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{clonality}
\alias{clonality}
\title{clonality}
\usage{
clonality(x, entropy = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector or tcr object}

\item{entropy}{Logical indicating if output should be Entropy (TRUE) or
clonality (FALSE), given by 1 - Entropy.}

\item{merge}{If x is a tcr object, the output will be a tcr object with the
clonality merged into the metadata.}
}
\value{
If x is a vector, the function returns the clonality. If x is a tcr
object, and merge=T, then the function returns a tcr object with clonality
included in the metadata for all samples.
}
\description{
Computes the clonality (normalized Shannon Entropy) of vector or all samples
in a tcr object
}
\author{
Alexander Hopkins
}
