% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_clone.R
\name{exp_clone}
\alias{exp_clone}
\title{exp_clone}
\usage{
exp_clone(x, y)
}
\arguments{
\item{x, y}{Vectors representing the receptor counts from two samples}
}
\value{
A vector of fisher test p values for each clone, the same length as x and y
}
\description{
Calculates p values for each clone using a fisher test.
}
\author{
Alexander Hopkins
}
