% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.DataFrame.R
\name{filter.DataFrame}
\alias{filter.DataFrame}
\title{filter.DataFrame}
\usage{
\method{filter}{DataFrame}(.data, ...)
}
\description{
Method for using dplyr filter on DataFrames
}
\author{
Alexander Hopkins
}
