% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{fisher}
\usage{
fisher(x, s)
}
\arguments{
\item{x}{A 1x2 matrix containing the counts of one clone (typically before
and after a treatment)}

\item{s}{A 1x2 matrix containing the sum of all receptors in each sample
clonality (FALSE), given by 1 - Entropy.}
}
\value{
A p value derived from a one sided fisher test (testing the
hypothesis that the clone expanded from the left sample to the right sample)
}
\description{
Performs a fisher test using a contingency table where the top row is the
abundance of one clone and the bottom row is the total counts of all other
clones in that sample
}
\author{
Alexander Hopkins
}
