\name{immunoSeqR-package}
\alias{immunoSeqR-package}
\alias{immunoSeqR}
\docType{package}
\title{
Tools for analyzing data from Adaptive's ImmunoSeq platform
}
\description{
Results from T cell receptor beta chain sequencing can be aggregated and manipulated using this package.
}
\details{
\tabular{ll}{
Package: \tab immunoSeqR\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-01-09\cr
License: \tab None (yet)?\cr
}
overview
}
\author{
Alexander Hopkins

Alexander Hopkins <achopkins@gmail.com>

}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
