% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_aggregate.R
\name{iseqr_aggregate}
\alias{iseqr_aggregate}
\title{iseqr_aggregate}
\usage{
iseqr_aggregate(ds)
}
\arguments{
\item{ds}{A TCR object made with iseqr_make_tcr.}
}
\value{
the aggregated dataset
}
\description{
Collapse synonymouse nucleotide sequences, summing the values in each sample
}
\author{
Alexander Hopkins
}
