% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_exp_cl2.R
\name{iseqr_exp_cl}
\alias{iseqr_exp_cl}
\title{iseqr_exp_cl}
\usage{
iseqr_exp_cl(ds, comps, category = "type", alpha = 0.05, min.count = 5,
  output = "tcr")
}
\arguments{
\item{ds}{A tcr object}

\item{comps}{A list of comparison types}

\item{category}{The field to look for 'comps'. Default: type}

\item{alpha}{The threshold over which clones with BH adjusted p-values will
be considered 'expanded'. Default: 0.05}

\item{min.count}{The minimum number of clones (sum of both samples) to be
included in the analysis, Default: 5}
}
\value{
If output=tcr, a tcr object with the number of expanded clones merge
into the dataset. If output='nums', a data.frame with the number of expanded
clones for each comparison. If output='p_vals', a matrix (ncol(assay(ds)) x
length(patients)) with the p-values for each clone in each comparison.
}
\description{
Calculates the number of clones that expand relative to a baseline sample in
a tcr experiment
}
\note{
It is not recommended (but not STRICTLY enforced) that you use the same
baseline sample for comparisons (i.e. the first elements of all items in
comps should be the same). If you REALLY know what you are doing, then you
could use multiple baseline samples.

If output='p_vals', then only one comparison can be made at a time
(lenght(comps) should be 1).
}
\author{
Alexander Hopkins
}
