% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_make_tcr.R
\name{iseqr_make_tcr}
\alias{iseqr_make_tcr}
\title{iseqr_make_tcr}
\usage{
iseqr_make_tcr(ds, dict, remove = T)
}
\arguments{
\item{ds}{A dataset containing Adaptive TCR data, made with `iseqr_merge`}

\item{dict}{A dictionary containing the sample metadata. This must contain a
field called `fn` with sample names matching the original tsv files (or, more
accurately the colnames(ds)).}

\item{remove}{Logical, indicating if non-productive clones and clones with
stop codons should be removed}
}
\value{
A `tcr` object
}
\description{
Make a tcr object from an imported adaptive dataset and a dictionary
containing sample metadata.
}
\author{
Alexander Hopkins
}
