% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_morisita.R
\name{iseqr_morisita}
\alias{iseqr_morisita}
\title{iseqr_morisita}
\usage{
iseqr_morisita(ds, comps, merge = T)
}
\arguments{
\item{ds}{A tcr object}

\item{comps}{A list of comparisons to be made}

\item{merge}{Logical indicating if the results should be returned alone, or
merged into the tcr object}
}
\value{
A data.frame containing the Morisita indicies (if merge=F) or a tcr
object with the Morisita indicies included in the metadata
}
\description{
Computes the Morisita index (similarity) across multiple comparisons in a tcr
object
}
\author{
Alexander Hopkins
}
