% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_order.R
\name{iseqr_order}
\alias{iseqr_order}
\title{iseqr_order}
\usage{
iseqr_order(plot_ds, ds, reorder = T)
}
\arguments{
\item{plot_ds}{A dictionary or plot_ds object}

\item{ds}{An object created with iseqr_merge}

\item{reorder}{Logical indicating if the plot_ds should be reordered}
}
\value{
If reorder=T, a plot_ds object in the same order as the ds. If
reorder=F, the order in which it should be placed to acheive this.
}
\description{
Checks that the dictionary and data are in the same order and, if necessary,
reorders the dictionary to fit the data. This is used in the iseqr_merge
function
}
\author{
Alexander Hopkins
}
