% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_plot_metrics.R
\name{iseqr_plot_metrics}
\alias{iseqr_plot_metrics}
\title{iseqr_plot_metrics}
\usage{
iseqr_plot_metrics(plot_ds, metric, by, type = NA, sm = TRUE,
  hjust = -0.5, labels = TRUE, ...)
}
\arguments{
\item{plot_ds}{A data.frame or tcr object to be plotted}

\item{metric}{The metric to be plotted (column name)}

\item{by}{The column to split the data by. This must be a numeric.}

\item{type}{A value to restric the data by (must correspond to a value in the
"type" field of plot_ds)}

\item{sm}{Logical indicating if trend line should be added with geom_smooth}

\item{hjust}{To be passed to geom_text}

\item{labels}{Logical indicating if labels in the "patient" field of plot_ds
should be printed}
}
\value{
A ggplot object graphing the data
}
\description{
Generated ggplot objects from a data.frame or tcr object
}
\author{
Alexander Hopkins
}
