% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_misc.R
\name{mean_only}
\alias{mean_only}
\title{mean_only}
\usage{
mean_only(x)
}
\arguments{
\item{x}{data to be summarized}
}
\description{
Computes the mean and outputs it for use with stat_summary function
in ggplot
}
\author{
Alexander Hopkins
}
