% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morisita.R
\name{morisita}
\alias{morisita}
\title{morisita}
\usage{
morisita(x, y)
}
\arguments{
\item{x, y}{Count vectors (integers) to be compared}
}
\value{
The Morisita distance between the vectors
}
\description{
Compute the Morisita distance between two vectors. This is a metric of
similarity between two populations.
}
\author{
Alexander Hopkins
}
