% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{overlap}
\usage{
overlap(x, y)
}
\arguments{
\item{x, y}{Count vectors (integers) to be compared}
}
\value{
The overlap between the vectors
}
\description{
Compute the overlap between two vectors. This is a metric of
similarity between two populations.
}
\author{
Alexander Hopkins
}
