\name{refactor}
\alias{refactor}
\title{Re-factors all columns of a data frame of class 'factor'}
\description{After subsetting a data frame, this function can be used to remove the unused 
factor levels from all the factor-class columns at once.}
\usage{df.r <- refactor(df)}
\arguments{
  \item{df}{The input data frame, containing (usually) multiple factor columns to be 
refactored}
}
\details{}
\value{
  \item{df.r}{The output data frame, with the extra levels removed}
}
\references{}
\author{
Alexander Hopkins
}

\seealso{}
\examples{}
\keyword{}
