% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{richness}
\alias{richness}
\title{richness}
\usage{
richness(x, ...)
}
\arguments{
\item{x}{A numeric vector or tcr object}

\item{merge}{If x is a tcr object, the output will be a tcr object with the
richness merged into the metadata.}
}
\value{
If x is a vector, the function returns the richness If x is a tcr
object, and merge=T, then the function returns a tcr object with richness
included in the metadata for all samples.
}
\description{
Computes the richness (total number of unique elements) in a sample
}
\author{
Alexander Hopkins
}
