\name{shared}
\alias{shared}
\title{Counts shared clones}
\description{Counts the number of clones shared by two samples}
\usage{num <- shared(x,y)}
\arguments{
  \item{x}{One sample ( a column of an immunoSeqR dataset, for example)}
  \item{y}{Another sample}
}
\details{}
\value{
  \item{num}{The number of shared clones}
}
\references{}
\author{
Alexander Hopkins
}
\note{}

\seealso{}
\examples{}
\keyword{}
