% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson.R
\name{simpson}
\alias{simpson}
\title{simpson}
\usage{
simpson(x)
}
\arguments{
\item{x}{An integer vector indicating the counts}
}
\value{
The Simpson Index
}
\description{
Computes the Simpson Index (a diversity measure) of a distribution. This is
used in the calculation of the Morisita Index.
}
\author{
Alexander Hopkins
}
