% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{size}
\usage{
size(x)
}
\arguments{
\item{x}{An R object}
}
\value{
The size of x in MB
}
\description{
A simple function to print the size of an R object, formatted in MB
}
\author{
Alexander Hopkins
}
