% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss2n.R
\name{ss2n}
\alias{ss2n}
\title{ss2n}
\usage{
ss2n(x, y)
}
\arguments{
\item{x, y}{Matricies of the same length and width >1}
}
\value{
A vector the same length as x and y with the signal to noise scores
}
\description{
Signal to noise score for two groups with replcates. This is defined as the difference in
means between groups divided by the sum of their standard deviations.
}
\author{
Alexander Hopkins
}
