% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\title{stars}
\usage{
stars(t)
}
\arguments{
\item{t}{A htest (or similar) object}
}
\value{
A character string containing stars representing the significance of
the p-values
}
\description{
Outputs stars to denote p-values in graphics outputs.
}
\author{
Alexander Hopkins
}
