% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonimity.R
\name{synonymity}
\alias{synonymity}
\title{synonymity}
\usage{
synonymity(x)
}
\arguments{
\item{x}{A character list with (probably) repeats}
}
\value{
A named list, the length of unique(x) counting each occurrence
}
\description{
Computes the synonymity (how many times each unique element appears). In
hindsight, this does exactly what 'table' does.
}
\author{
Alexander Hopkins
}
