% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcr.R
\docType{class}
\name{tcr-class}
\alias{tcr-class}
\alias{tcr}
\title{tcr class objects and basic methods}
\value{
A tcr class object
}
\description{
This is the constructor function for a tcr class object, which extends
SummarizedExperiment
}
