% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\title{Total sequences}
\usage{
total(x, ...)
}
\arguments{
\item{x}{A numeric vector or tcr object}

\item{merge}{If x is a tcr object, the output will be a tcr object with the
richness merged into the metadata.}
}
\value{
If x is a vector, the function returns the sum. If x is a tcr
object, and merge=T, then the function returns a tcr object with
Total.Sequences included in the metadata for all samples.
}
\description{
Computes the sum of a vector
}
\author{
Alexander Hopkins
}
