% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_stats.R
\name{delta_stats}
\alias{delta_stats}
\title{delta_stats}
\usage{
delta_stats(ds, comps, metric, merge = T)
}
\arguments{
\item{ds}{A tcr object}

\item{comps}{A list of comparison types}

\item{metric}{A character specifying the name of the field to be used for the
change calculation}

\item{merge}{If x is a tcr object, the output will be a tcr object with the
clonality merged into the metadata.}
}
\value{
If merge=T, a tcr object with 'Log2.Fold.Change.in[metric]' included
in the metadata. If merge=F, a data.frame containing the fold changes.
}
\description{
Computes the change in metrics and (optionally) merges the results back into
a tcr object
}
\author{
Alexander Hopkins
}
