% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iseqr_merge.R
\name{iseqr_merge}
\alias{iseqr_merge}
\title{iseqr_merge}
\usage{
iseqr_merge(all_files, data = "estimatedNumberGenomes",
  nucleotide = "nucleotide", aminoAcid = "aminoAcid")
}
\arguments{
\item{all_files}{path to directory containing tsvs}

\item{nucleotide}{the column header in the tsv denoting the nucleotide
sequence}

\item{aminoAcid}{the column header in the tsv denoting the amino acid
sequence}
}
\value{
a data.frame with all the data joined together
}
\description{
Merge Adaptive ImmunoSeq tcr samples into a data.frame
}
\author{
Alexander Hopkins
}
