% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_misc.R
\name{mean_95}
\alias{mean_95}
\title{mean_95}
\usage{
mean_95(x)
}
\arguments{
\item{x}{data to be summarized}
}
\description{
Computes the mean and 95% confidence intercal. For use with stat_summary function
in ggplot
}
\author{
Alexander Hopkins
}
