% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_misc.R
\name{mean_sd}
\alias{mean_sd}
\title{mean_sd}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{data to be summarized}
}
\description{
Computes the mean +/- standard deviation. For use with stat_summary function
in ggplot
}
\author{
Alexander Hopkins
}
