% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{shared}
\alias{shared}
\title{shared}
\usage{
shared(x, y)
}
\arguments{
\item{x, y}{Numeric vectors of the counts of clones in each sample}
}
\value{
The number of clones that are present in both samples
}
\description{
Computes the number of clones that are present in two samples.
}
\author{
Alexander Hopkins
}
