% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tert.R
\name{tert}
\alias{tert}
\title{tert}
\usage{
tert(x)
}
\arguments{
\item{x}{A distribution}
}
\value{
A vector indicating which tertile each data point falls in
}
\description{
Computes the tertiles for a distribution
}
\author{
Alexander Hopkins
}
